rearrange grequests, prepare for selenium fix
